Import HDZero OSD Driver from iNav
Mostly from https://github.com/iNavFlight/inav/pull/7668/files commit
Notable changes:
* And added missing functions to bitarray.h/.c, msp_serial.h/.c
* Used "drivers/max7456_symbols.h" instead of "drivers/osd_symbols.h"
* Used baudRates[portConfig->msp_baudrateIndex] instead of
baudRates[portConfig->peripheral_baudrateIndex]

Additional changes:
serial: serialPortFunction_e now needs 32 bits

How to use:

In CLI, Set FUNCTION_HDZERO_OSD to the appropriate serial port:
Eg. Assuming to use serial port 4, enter the following CLI commands to
enable FUNCTION_HDZERO_OSD in EmuFlight Configurator 3.0.1:
    serial 3 65536 115200 115200 0 115200
    save

+++ src/test/unit/telemetry_ibus_unittest.cc
@@ -185,7 +185,7 @@ bool telemetryDetermineEnabledState(portSharing_e portSharing)
 
 
 bool isSerialPortShared(const serialPortConfig_t *portConfig,
-                        uint16_t functionMask,
+                        uint32_t functionMask,
                         serialPortFunction_e sharedWithFunction)
 {
     EXPECT_EQ(findSerialPortConfig_stub_retval, portConfig);
